+++ atom.portable/tools/chocolateyUninstall.ps1
@@ -5,9 +5,14 @@ $shortcutPath = Join-Path $([Environment]::GetFolderPath([System.Environment+Spe
 Remove-Item $shortcutPath -ErrorAction SilentlyContinue -Force
 
 $logPath = Join-Path $Env:ChocolateyPackageFolder 'atom.txt'
+Write-Verbose "Reading Installation Path from $logPath"
 $installationPath = Get-Content $logPath
 Write-Verbose "Previous Installation Path: $installationPath"
 
 Uninstall-BinFile -Name 'atom'
 
-Remove-Item -Path $installationPath -ErrorAction SilentlyContinue -Recurse -Force
+if (Test-Path $installationPath -ErrorAction SilentlyContinue) {
+    Remove-Item -Path $installationPath -ErrorAction SilentlyContinue -Recurse -Force
+} else {
+    Write-Warning "Installation Path missing or not found: $installationPath"
+}
